Move tests/compiler to compiler/test
