+++ pom.xml
@@ -57,6 +57,7 @@ limitations under the License.
         <module>compiler/frontend/beam</module>
         <module>compiler/frontend/spark</module>
         <module>compiler/optimizer</module>
+        <module>compiler/test</module>
         <module>examples/beam</module>
         <module>examples/spark</module>
         <module>runtime/common</module>
